escudo-ia-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-nmap = "^0.7.1"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
pydantic-settings = "^2.10.1"
python-multipart = "^0.0.20"
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
trio = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
